refactor(client):refactor and update styles for the sign in flow
