+++ packages/amplication-client/src/User/Login.scss
@@ -9,11 +9,8 @@ $message-margin-bottom: 24px;
     font-size: 24px;
     font-weight: 600;
   }
-  &__sso {
-    text-decoration: underline;
-    color: var(--gray-base);
-    font-size: 12px;
-    font-weight: 400;
+  &__sso-link {
+    margin-bottom: 0;
   }
 
   &__signup {
@@ -46,27 +43,6 @@ $message-margin-bottom: 24px;
     margin-bottom: $message-margin-bottom;
   }
 
-  &__or {
-    display: flex;
-    flex-direction: row;
-    & span {
-      padding: 0 8px;
-      color: var(--gray-50);
-    }
-  }
-  &__or:before,
-  &__or:after {
-    content: "";
-    flex: 1 1;
-    border-bottom: 1px solid var(--gray-50);
-    margin: auto;
-  }
-
-  &__work_email_btn {
-    height: 34px;
-    border-radius: 4px;
-  }
-
   &.amp-button--primary[disabled] {
     background-color: #2c3249;
   }
@@ -74,34 +50,6 @@ $message-margin-bottom: 24px;
   &.amp-button {
     margin-top: 0px;
     border-radius: 4px;
-  }
-
-  &__reset_message {
-    border: 1px solid var(--theme-green);
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    border-radius: 4px;
-    color: var(--gray-20);
-    padding: 7px;
-    & i {
-      align-self: center;
-      margin-right: 10px;
-    }
-  }
-
-  &__work_email {
-    & .amplication-label__ {
-      color: var(--gray-20);
-      font-size: 10px;
-    }
-    & input {
-      max-width: 100%;
-    }
-    & input:hover:not(:disabled),
-    input:focus:not(:disabled) {
-      border: 1px solid var(--gray-base);
-    }
   }
 
   &__title {
