+++ packages/amplication-client/src/User/constants.ts
@@ -1,6 +1,5 @@
 import { PageContent } from "../Layout/WelcomePage";
 import amplicationLogo from "../assets/logo-amplication-white.svg";
-import recastLogo from "../assets/logo-recast.svg";
 
 type PageContentOptions = {
   [key: string]: PageContent;
@@ -9,13 +8,6 @@ type PageContentOptions = {
 export const DEFAULT_PAGE_SOURCE = "default";
 
 export const SIGN_IN_PAGE_CONTENT: PageContentOptions = {
-  rc: {
-    name: "Recast",
-    title: "Convert your excel sheet into Node.js resource",
-    subTitle: "",
-    logo: recastLogo,
-    message: "Recast is built on and powered by Amplication.",
-  },
   [DEFAULT_PAGE_SOURCE]: {
     name: "Amplication",
     title: "Instantly generate quality Node.js services",
